Add NotificationShelfController.
This moves injection from NotificationShelf into its controller,
creating a subcomponent for the view.

This is a relatively light-weight version of this change. Very little
logic is actually moved into the controller, with a handful of public
methods on the controller simply acting as proxies to the view for
now.

Bug: 147245740
Test: manual && atest SystemUITests
Change-Id: Ic74dc6a2ef83bcc2e1a86080d77b156b2b3e8c15

Former-commit-id: f152f9d2ab96b47593875e7d6a2e932587ddf904