Restart the WebView zygote after a WebView update.
The zygote is stopped when a WebView update occurs, and will be
automatically restarted the first time an app tries to use it, but this
means the first app to use WebView after an update pays a fairly large
startup time penalty.

Instead, queue a background task to wait for RELRO creation to be
finished and then start the zygote immediately afterward, as we do at
boot time, saving the app from paying this cost.

Fixes: 112485729
Test: atest WebViewUpdateServiceTest
Change-Id: I14159980851f423675b6a4ec0905cecda63e0e54

Former-commit-id: bc8fbe40d2fc2ee350d23d45610d5e65b24707d5