Override the label of the SIM Toolkit application dynamically
The list of the menu items provided by SET UP MENU command shall be a
part of the menu system of the terminal and the alpha identifier shall
be used for the title of the menu in accordance with the clause 6.4.8
SET UP MENU in ETSI 102.223. The terminal shall be able to display the
alpha identifier as the lable of the launcher activity of SIM Toolkit.

Bug: 113638339
Test: Confirmed the label of SIM Toolkit can be controlled as expected.

Change-Id: I65f6d4d9247bf96985c3b6e8697cdb641b6e3306

Former-commit-id: 96c37d26cb1604aee8b76fe84918911932d36364