Remove precision loss on brightness slider.
Replacing this check with an equivalent float one allows all backlight values to be used. Previously, each slider change was being ignored if the int values were equivalent. Int values have a range of 1 to 255, however, the hardware can have more precision than this, so its more precise to do the comparison using floats, not ints.

Bug: 157454239

Test: Manual, `ROBOTEST_FILTER=BrightnessLevelPreferenceControllerTest m -j RunSettingsRoboTests`, `cat /sys/class/backlight/panel0-backlight/brightness`.

Change-Id: I15eab92b465652fc0024ed92045cc731492d88ba

Former-commit-id: 3dfe3448c0f4ef48dfeab1c8178e4a0736c10f9b