Non-system camera PPA set by DPC
Background
ag/10760163 introduced a change which requires
camera intents to match system apps only.
As a result, the DPC will no longer have the
aility to control the camera PPA.

This CL gives the DPC the ability to set a
camera PPA, by adding a new field isSetByDPM
to the PersistentPreferredActivity. If this
field is set, the camera intent does not have to
match a system app.

Changes
* Added isSetByDPM to PersistentPreferredActivity
* Added helper method in PackageManagerService to
  check if an intent is an image capture intent
  and not set by the DPC
* updated the is isImplicitImageCaptureIntentAndNotSetByDpc
  parameter to updateFlagsForResolve() in
  PackageManagerService

Bug: 153921883
Test: atest com.android.server.pm.PackageManagerServiceTest
      atest com.android.cts.devicepolicy.MixedDeviceOwnerTest#testPersistentIntentResolving
      atest com.android.cts.devicepolicy.MixedProfileOwnerTest#testPersistentIntentResolving
      atest com.android.cts.devicepolicy.MixedManagedProfileOwnerTest#testPersistentIntentResolving
Change-Id: I667a7e613b76514ecc1d814393d173c2987ca0b6

Former-commit-id: 1ab58c3b8b8aac1645103a491515c3d3efeec156