Remove IME surface when occured in multi-window mode
On split-screen, IME is not controlled by the client and it's critical
to remove the surface when hiding the IME.

DisplayImeController doesn't have the access to the actual enabled
InputMethodManager. Instead, this patch introduced a new interface in
InputMethodManagerService to let system UI remove the currently enabled
IME surface when needed.

Bug: 155660756
Test: go/wm-smoke
Test: See reproduce steps in the bug.
Change-Id: I800dbab20aa90fd549ea1c07da85572311ecb7b3

Former-commit-id: 460dabda46d543b6db9c963bd4f85b89575e08d8