DO NOT MERGE Fix work profile change race condition in LMS
It's possible for LMS to find out about work profile changes through
external clients before LMS is notified via the UserInfoHelper. This
means LMS will have stale information on work profiles, and will not
refresh that information when appropriate. We add notifications for
profile users as well so that LMS will refresh appropriately.

Bug: 157444953
Test: manual
Change-Id: I4dc9de6019847b790f2d4aa90b66484cf0a14674

Former-commit-id: 9b0b32bd17d25ba6763139ce71f6d07240432e6a