Revert "Persist implicit overlay configurator actor policy"
This reverts commit f2e3b73aaedc9a7280a7aaa34effb32fec0a409d [formerly af93f5de66f09797345d2171306c9de2951bf85c].

The overlays.xml file version should not be reverted to 3.
Increasing the version caused b/158054824 so for devices that have not
upgraded to version 4, do not ignore a version 3 overlay file when
restoring settings. In a future change, we must create testing so that
version upgrades always migrate old overlays.xml files to new
overlays.xml files.

Bug: 157266239
Bug: 158054824
Change-Id: I3f8c49d30a1961a756bbed9b2ab01aeb1f887e55

Former-commit-id: e88ef74249076251c28bf55be58943f325d2a1f8