Force-clear nav bar flags when phone finishes booting.
It looks like there may be a bug in StatusBarManager that prevents it from correctly hiding the home button
if flags to show the home button have not been sent since a restart.

This side-steps that problem by sending an empty dismiss call once after the phone restarts, then following it
immediately with a call to hide the home/recents buttons as needed.

Test: Manual -- no home button visible after restarting phone and opening power menu on lock screen.

Bug: 155663717
Change-Id: I489472e364b34390d20baa28d838687498abe69c

Former-commit-id: 5d9a184ea49f1add8dd456241beff8bec3b4e8db