Controls UI - Correctly persist seeding state
It is incorrect to attempt to modify a Set that is returned from
SharedPreferences.getStringSet(). Won't work. Create a new Set.

Also handle the case on device restore, where we need to set the
seeded flag when controls exist.

Also fix the uninstall logic to look at the actual service list and
not the favorites list

Fixes: 158678748
Test: seed, remove all controls, and restart phone
Change-Id: I3dbc1c7cbac999b06b572b8348cf41617a73adde

Former-commit-id: 9da55a7a38b5bf4b5731fa8ef6fb61e1398e5322