Expose isServiceMultiuser
Expose ServiceInfo.isServiceMultiuser so the server can take different
actions on bind / unbind.

For example: when a client runs as the current user there are races
between processes when the user switches. i.e. the system server is
aware that the current user is switching from user A to user B
"immediately", but the client running as user A won't be shut down for a
while and can end up being rebound and sending events for user A. Any
binder flapping associated with the user A -> user B switch should
generally be ignored for a period of time.

In the case of a client service that always runs as the system user, the
same binder behavior is unexepected would be a more serious error and
should be acted on immediately or reported as a configuration error.

Test: build only
Change-Id: I18f6629aac18076af82b4f6fbf08c4fe5d1c1fd8

Former-commit-id: a5357c95207ed6c19e89d03a5ad03cf9b81addd5