Remove incorrect isAppUid check reading memory state
The check misclassifies processes as apps (it only checks lower bound,
whereas there's also an upper bound of 20000-1).

At the same time it is redundant. It was originally introduced to
quickly filter out processes that were known to AM, but nowadays we
explicitly remove these from the set of processes to scan.

Bug: 158470048
Test: manual test, android.cts.statsd.atom.UidAtomTests#testProcessMemorySnapshot, 'android.cts.statsd.atom.UidAtomTests#testProcessMemoryHighWaterMark
Change-Id: Ie24107ccea31efae40277d8d46fa36f978f2481e

Former-commit-id: e0ec1f3ce4c2b3e23be611e43f50403cef11ecaa