Leave Prox registered when safe to do so.
This CL allows a proximity sensor operating in dual mode to
leave its secondary sensor regitered where it is told it is
safe to do so. This is generally when the screen is in some
sort of doze or aod mode. The sensor does not itself manage
this state but expects clients to tell it when it it safe.

Fixes: 157717162
Test: atest SystemUITests && manual
Change-Id: Ie89e80aa7440b4964fb61c144ae3b8bc4808f0ac

Former-commit-id: 2a7d804124deecb6aaf98a60b6dd85b4b7fe42d9