Don't lock rotation to natural rotation when entering overview
Overview (recents) will use fixed portrait orientation if home
rotation is disabled, then the orientation will be portrait when
leaving overview to a application which was rotated to landscape.
Which is inconsistent from user perspective.

This change ignores the rotation change caused by the overview
once. So the original user rotation can be preserved.

Bug: 148982179
Test: 1. Disable auto-rotation and home rotation.
      2. Launch an activity in portrait.
      3. Rotate the device to landscape and press rotation button.
      4. Enter overview (recents).
      5. Return to the application.
      Check with 2-button, 3-button, and gesture navigation.
      The device should be in landscape after step 5.
      And the behavior doesn't change with home rotation enabled.

Change-Id: I43005753031b13e66c3c388b221d51ce3fd543bf

Former-commit-id: 95cb52a647c512b77aa7fd13081cecd86cab1a52