Close assistant activities on closeSystemDialogs
When the assistant is configured to be on top of the dream, it will have
higher z-order than any other activity. If it is also opaque, it will
prevent other activities from starting. We want to close the assistant
on closeSystemDialogs to allow other activities to start, e.g. on home
button press

Bug: 158117282

Test: atest AssistantStackTests
Change-Id: I226d95b82c1f5df5e5327ef85859d9b137d6e0e7

Former-commit-id: 72d581e60281a1dbfb8d411c783e8839f247dfa5