Check for null service in ScreenshotHelper
We were testing whether mScreenshotConnection was null as a proxy
for whether we had a running screenshot service. In principle
mScreenshotService should never be null if mScreenshotConnection
is not null, but it's safer to check. This may fix the crash noticed in
early May (attached bug), and either way is better for the stability
of the process.

Bug: 156366251
Fix: 156366251
Test: manual
Change-Id: Iea669a5a2ca2334760141ec6fec8ec7ffb01f25c

Former-commit-id: 33aff26285027251ef5f95d9b40e17d1eb52eeef