Respect "Don't add users on lock screen" regardless of lockscreeen secure state.
Currently this system setting only applies when a secure lock screen is
set. It's more consistent to apply the preference whether the lock
screen is secure or not (even if in the case of an insecure lockscreen
there is nothing to stop someone "unlocking" the device and then adding
more users through system or quick settings.

Bug: 152587665
Test: Manual
Change-Id: I8281e7bea291a77a2eec7aaf47e8531bb3925c10

Former-commit-id: ad196f9316e3279cf35f9d67a50e82af173828ac