Dismiss dream with launchTaskBehind
When we are launching underneath an AOT screensaver, force launch-behind,
attach the process, and go through the resume lifecycle events instead
of creating the activity in stopped state.

The check for turning the window on during relayout also needs a fix to
check for isDreaming() as well as isInteractive() when deciding whether
to call PowerManager.wakeUp() during window relayout.

Test: atest DreamManagerServiceTests
Bug: 154290545
Bug: 159821769
Change-Id: Ib3f6e6caad3abd793f5c66eece8beffcffde2017

Former-commit-id: 0e8e167e41d9a203efb23af8f211408142da906d