Call finishDrawing and notify when Window is removed or hidden
There are cases where a sync transaction is called on a window that's
going to be removed or hidden. In that case, it will never get a
finishDrawing and possibly not get a prepareSurfaces call. Since we're
not really waiting for anything, we can call finishDrawing and
immediately call notifyBlastSyncTransaction.

This change also moves the blast sync timeout removal into
notifyBlastSyncTransaction in case there's another issue where
finishDrawing is called but not notifyBlastSyncTransaction. The reason
for this is the device can at least go back to a normal state instead of
permanently staying stuck.

Fixes: 159507259
Test: WindowOrganizerTests
Change-Id: If483e3769a98cc100c86893f2132aced5f064fe6

Former-commit-id: 355befde244080023631eba3f553471f05de6fd4