Require explicit `android:exported` in manifests for services
Starting S+, if an intent filter is defined, we are requiring that the
android:exported flag also be explicitly defined in the manifest. With
this change apps targeting S+  that do not meet this requirement will
fail to install.

There is no change in the default behavior of the `android:exported`
flag when there is no intent filter defined (i.e. the flag defaults to
false in these cases).

Bug: 150232615
Test: Flashed the build on a device and tested with test apps
Change-Id: Icc34b533cd7420e5754f7a1e7ae9fde937febf7d

Former-commit-id: 553bada3296ec8ffb8f15ddca1bb295a40e9b366