Transfer IME target right away if old target was home
In most cases we want to let the old window be the IME target until it's
finished closing. That is to prevent flicker when switching between
windows that both show the IME. However, when home is the old IME target,
we want to immediately change the target to the new window since the
app is opening.

Test: Open app from home that opens IME when launched
Fixes: 159846344
Change-Id: Ib826b53415b1b1affde76e609d999e1b6f2d2b79

Former-commit-id: 0c1a5237fb0ff1b1c7984fcc3926ca3594e737e0