Fix issue where PIP jumps after being tapped after being dragged through the dismiss target.
onBoundsAnimationEnd does not set the permanent bounds if mSpringingToTouch=true, since that indicates that PIP caught up with the touch location - and the touch location is a temporary position. We eventually set the permanent bounds after flingToSnapTarget's animation ends.

However, if PIP was flung to a snap target while mSpringingToTouch=true, those permanent bounds were never set. Starting a drag gesture would have set the permanent bounds, which is why this is only visible if you tap immediately after doing all this.

Bug: 159909223
Test: tap pip after dragging it through dismiss without moving it again
Change-Id: I67a9c19bf8dafab4cee646539d9f5a4575a14ffd

Former-commit-id: 40d313a0462a1d97a5d16ca3f40d3faeb62bc916