Do not put in-recent task in hidden task list
Tasks that cannot be accessed recently are hidden tasks, and the tasks
will be removed when going to home to avoid leakage. But the task may
be added to recent task again according to back stack, then it should
not be hidden anymore, otherwise it may be removed unexpectedly.

Also guard the case when performing the removal of hidden tasks.

Fixes: 149538962
Test: RecentTasksTest#testAddTasksHomeClearUntrackedTasks_expectFinish
Change-Id: I3de151aaeaad311f863c3867102a916779370407

Former-commit-id: a9391ebe84bf74ef184dd2cdf1e1f22d118551bb