Promote NO_ISOLATED_STORAGE app op to SystemApi(MODULE_LIBRARIES)
This app op was introduced to support --no-isolated-storage
instrumentation flag in MediaProvider mainline module. Because it was
very late in the release cycle, app op wasn't exposed as @SystemApi in
rvc-dev, even though there CTS tests written around it, and
MediaProvider module depends on it.

To make things fair, expose it as @SystemApi.

Test: atest CtsScopedStorageHostTest
Bug: 159708202
Change-Id: Ifddf2cb920285eda5f859972937b0f9084fc32a3

Former-commit-id: 89f7de19c3d3a702545371779d4d706c1aaa9241