Fixed that texts of players would often remain empty
Because we were binding the view too late, at the time
we were taking our measurements, they were not bound
yet and hence the measurements were completely wrong.
In the past this worked accidentally, since they sizes
of the views didn't really depend on the content, but
various changes have been using wrap_content recently,
which could result in empty titles and songs.
Similarly, buttons could not have the right state
as a result and would be positioned wrongly.

This also could lead to the players having the wrong
order since isPlaying is depending on the view being
bound.

Fixes: 160197584
Test: add resumable player, observe no empty texts
Change-Id: I8fe4fb683efdabea4f1e985515f4c31d610170ea

Former-commit-id: a0447977791f9cf2bb24bd6b54ea0c58f557d8ab