Collapse shade on ACTION_CANCEL (sometimes)
As of ag/11546928, the framework emits an ACTION_CANCEL when the
proximity sensor turns off the screen, not an ACTION_UP. Previously, we
always expanded the shade in response to an ACTION_CANCEL. However,
after this change, this caused Dialer users to end up with an expanded
shade after putting the phone to their face, making it easy to
face-dial their notifications. This CL changes the behavior to put the
shade back to where it was when the gesture started -- either expanded
or closed.

Fixes: 155362798
Test: manual
Change-Id: I2d11d9f03bfe1e905028a640e2eb871d32e37248

Former-commit-id: 7c6dd3d568c028d197322ad148d109c745ec53f8