Fixes component-based visibility recompute
This change fixes a bug that was resulting in any forceQueryable
package's queries->intent tags to be ignored. The logic in the
recomputeComponentVisibility was skipping if the source was force
queryable instead of the target. This change fixes that and adds the
correct check for the source (has QUERY_ALL_PACKAGES)

Test: atest AppEnumerationTests AppsFilterTest
Fixes: 159626154
Change-Id: Iaba1c52e6d645db43715642e7db9eaacdfeefb7d

Former-commit-id: 76aa2deba7ae062787ba371a5005ea6118efcecd