Finish supporting floating-IME alpha in split-screen
There are a couple situations that cause floating IME:
zero-sized frame due to fullscreenIme and the IME itself
providing a frame that doesn't generate insets.

This CL adds detection for an IME frame that doesn't provide
insets and it also allows the subsidiary IME processers (like
divider) to disable alpha animation. This is necessary because
when divider adjust for ime, there is no content behind the
ime, so we can't have it fade in/out; however, if the divider
doesn't adjust, we want the fade in/out so it matches the
ime behavior when not in split.

Bug: 160915523
Test: Use fullscreen IME as well as making gboard floating.
      Open ime in either top or bottom for both cases.
Change-Id: If6fd192c6e7f843a958d98bf40c4d0866d818394

Former-commit-id: 46e68011e76b0f0dce4f80dbc04490f26ef0e1f3