Move task to back when the back button is pressed on root task
This CL changes the behavior of pressing the back button on a root
activity so that the task is moved to the back, instead of the activity
being finished.

This removes the IRequestFinishCallback which IActivityTaskManager would
previously have triggered to make the Activity finish itself.

Bug: 146418616
Test: atest WmTests
Test: manual: flash crosshatch, press back button on a root activity,
observe task is moved to back.

Change-Id: I48ec35f841ab3b306fe80845150000c390908f5e

Former-commit-id: 2dec42fa6678527fa6f99e1d897b019db6e59989