Don't auto-restart face auth after successful auth
Currently, face auth sends a message to restart face authentication
500ms after a successful authentication attempt. This is a holdover from
fingerprint and is unnecessary for face, since the latter is entirely
interrupt-driven. It is also leading to issues where face auth will scan
repeatedly when SystemUI freezes or is slow to respond. Thus, this
commit removes the restart behavior after the user has authenticated.

Test: atest com.android.keyguard
Test: Manually verified face auth functionality in Keyguard.

Bug: 157630051
Change-Id: I8bfb7870d3c68f6d40fdbf7209446782c5805f8f

Former-commit-id: b46f97157356ab27224e70cd8184f2d5cfb24721