Add dismissUserStats to NEM#performRemoveNotif
Now both notification pipelines require DismissedByUserStats when
removing a notification.

Also, refactored the DismissRunnable that gets called whenever a
notification is manually swiped away and dismissed from the
NotificationShade. Now, ExpandableNotificationRowController injects a
OnDismissCallback with a #onDismiss method that will get called whenever
a notification is manually swiped by the user OR is clicked with the
AUTO_CANCEL flag. Since it's injected it's easier to switch between the
new and old pipeline's OnDismissCallback (one which interacts with the
old pipeline's NotificationEntryManager, and another with the new
NotifCollection). Now this dismiss runnable doesn't need to be passed
around starting from inflation.

Test: atest SystemUITests
Change-Id: Iec985ce2c502462ee35cf86c1e0332168e578823

Former-commit-id: 189d4af72116f179773b38a9a922e11a74293c2f