Don't show new groups until their children inflate
At the moment, the pipeline has an awkward behavior when inflating
incoming notification groups:

1. Group is posted, we start inflating the summary and the children
2. Summary finishes inflating. Because it has no children, we just show
   it as a singleton notification in the shade.
3. First child finishes inflating. Now we have a group with one child,
   so we remove the summary and just show the child as a singleton.
4. Second child finishes inflating. Now we have a group with two
children, so we remove the singleton child, reattach the summary, and
add the children inside it.

This causes a lot of confusion for the NSSL. Instead, we should block
the ENTIRE group from being in the shade until its children are
completely inflated (with some timeout if it takes too long).

Test: manual
Test: atest SystemUITests
Change-Id: If9b31b3b07cae3abdb7b415ab6a090458bd29369

Former-commit-id: c396958240416e370b54f75ac288ec4a4305f3d0