Allow to replace fixed rotation state
Assume there are 2 activities with different fixed rotation states
A and B. And wallpaper was associated with A. When the wallpaper
target is changed to B, wallpaper should be able to change the
association to B so their rotation transform can be updated at the
same time. This also avoids that if the previous associated token
is somehow inactive, wallpaper won't in a dangling rotated state.

Also add a check of updating rotated launching app for the case:
launching a portrait app from a landscape app and trigger recents
animation immediately before the animation of portrait app is done.
And then finish recents animation by keeping the portrait app as
the top activity. The expected result should be that the display is
rotated from landscape to portrait seamlessly. But if the recents
activity is set to the rotated launching app, the display rotation
will be unable to update because when receiving animation done of
the portrait app, the recents animation is still active that skips
the rotation change. Then the portrait app will be updated to
landscape temporally.

Fixes: 161056612
Test: atest WindowTokenTests#testFinishFixedRotationTransform
Change-Id: Ic134c4326db836e35385f290f996f0d841da693d

Former-commit-id: 819e1209cdd873a4eb7402417a1cabd9e0039f6c