AudioService: enable reusing an AudioPolicy
Enable registering/unregistering an AudioPolicy, and then
registering it again later.
Doing so was failing before as when a policy was re-registered,
its mix count didn't restart at 0 on the client side, causing
the mix to have an invalid registration string not in sync with
the policy configuration on the server side.

Bug: 161491896
Test: atest AudioHostTest#testAudioPolicyReuse
Change-Id: I51fc41c46817791de4ce3a4f9d36c38d9466602e

Former-commit-id: f5f3e6894713fc0052c69d8f7d7b5b714a860ac7