Rebuild actor map if it's queried before onSystemReady
If a package is installed shortly after initial boot, it's possible
to query app filtering before the onSystemReady callback.

Instead of throwing an exception, this makes it just rebuild the map
on demand. This may result in a slight delay with the first initial
query, but that will be the extent of the performance impact.

Scanning should still prevent querying until the mapping is ready,
which avoids large rebuliding costs. Effectively the map is ready
to be built in this edge case, but the callback isn't instant.

Bug: 161816449

Test: atest com.android.server.om.OverlayReferenceMapperTests

Change-Id: I1c815cdc037e297198294b1517805ffebc759e3a

Former-commit-id: 50fff3b2b39f57fe020b80bb8555b093e45a50d2