Use vibrator HAL controller for prebaked effect
The new controller always triggers the callback, even if the HAL does
not support it. This requires changes in the VibratorService.java code
to rely on the native callback and use the ASYNC_TIMEOUT_MULTIPLIER for
the fallback.

This is the last step on the vibrator controller integration, and thus
removed the unused code from native VibratorService.cpp.

Bug: b/153418251
Test: atest FrameworksServicesTests:VibratorServiceTest
Change-Id: I65921dac3758a4f53612928a593af7461566a024

Former-commit-id: c45394fdb5f2681bc656d68f9ffb596d4f74b627