Increase parsing comparison test timeouts
The packageInfoEquality test case is hitting test timeout issues with
the default 60000 ms timeout.

The actual timeout reproduced during local testing was 56s, but that
seems close enough to the default 60s that performance differences
may account for the flakiness.

This uses a conservative 4 minute timeout, which accounts for potential
scaling with additional packages in newer releases.

Test: atest com.android.server.pm.parsing.AndroidPackageParsingEquivalenceTest

Change-Id: I63b2a768b69e0a4c264aa96ca3a3803c7ae9ff1d

Former-commit-id: 39779eb324c7f70841485771bf165b9570409c5c