setCurrentState(DESTROYED) called from main thread
Also address timing issues with bouncing calls between 2 handlers. In
particular, make sure that after a tile is DESTROYED, the state doesn't
change anymore due to leftover refreshState calls.

This sometimes would be caused as the result of an earlier tile added as
a callback on something that replies on add.

Test: atest TileServiceTest BooleanTileServiceTest
Test: manual, open QSCustomizer that destroys tiles
Test: atest com.android.systemui.qs
Change-Id: Idd4dc21558c6a851132dae7d29fc383b0258922d

Former-commit-id: 466687dec94db39a3c97846d72362f66cbf9db5d