Metrics no longer set whether condition is sliced
Previously, all conditions with a metric condition link (as well as the
primary condition for the metric) would have mSliced = true. The reason
for this was historical, since conditions did not initially have
dimensions, and the dimensions were inferred from the condition links.

However, now conditions have dimensions explicitly stated and the rules
(which have been the existing implementation are):
1. A simple condition is sliced if it has dimensions.
2. A combination condition is sliced if any of its children have
dimensions.

The reason for doing this is to remove all dependencies that metrics can
have on conditions to simplify implementing partial config updates.

Test: atest statsd_test
Bug: 162322841
Change-Id: Ic5bc77530682272fab0cb3b18484338a257043cd

Former-commit-id: 1d59325c01770c84b25419d315dc4d5d8f66a6a6