Change condition init to return current condition
To allow initialization of conditions to work for both config creation
and update, have init always fill the conditionCache with the current
condition.

For sliced dimensions, it should compute an OR over all dimensions,
returning true if any of the dimensions are true.

We can use this for setting the initial value of new combination
conditions that depend on existing initial conditions, or for new
metrics that depend on existing conditions.

Test: atest statsd_test. Existing tests pass, will add more for the
update path in the update for condition cl.
Bug: 162322841

Change-Id: Ib48479269828f2e071c5c46aba1129d52d4725de

Former-commit-id: cfdf67144c24f8ccbb30d883111f6bf105db6084