Clean up staging directories of abandoned sessions (4/n)
Now ag/12369403 prevents a staged session from being a child of a
non-staged session which however discloses a bug where the staging
directory is not correctly cleaned up when a non-committed staged
session is abandoned.

This CL fixes that and adds a test to check staging directories are
correctly deleted when abandoning both committed and non-committed
sessions.

Bug: 163976510
Test: atest StagedInstallTest StagedInstallInternalTest
Change-Id: I5d0eed9c5cc71c29a12e61471b2f870140899087

Former-commit-id: ec050f6ad4e6d40e77947d8b4cc1e5e4de5450af